LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/preload/modulepreload-expected.txt

CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin. Status code: 500
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin. Status code: 500
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin. Status code: 500

PASS link rel=modulepreload
PASS same-origin link rel=modulepreload crossorigin=anonymous
PASS same-origin link rel=modulepreload crossorigin=use-credentials
FAIL cross-origin link rel=modulepreload promise_test: Unhandled rejection with value: object "[object Event]"
FAIL cross-origin link rel=modulepreload crossorigin=anonymous promise_test: Unhandled rejection with value: object "[object Event]"
FAIL cross-origin link rel=modulepreload crossorigin=use-credentials promise_test: Unhandled rejection with value: object "[object Event]"
PASS link rel=modulepreload with submodules
PASS link rel=modulepreload for a module with syntax error
PASS link rel=modulepreload for a module with network error
PASS link rel=modulepreload with bad href attribute
PASS link rel=modulepreload as=script
PASS link rel=modulepreload with invalid as= value
PASS link rel=modulepreload with integrity match
PASS link rel=modulepreload with integrity match2
FAIL link rel=modulepreload with integrity mismatch promise_test: Unhandled rejection with value: object "[object Event]"
